requirements.txt
Flask>=2.3,<4
pywebview>=4.4
openpyxl>=3.1
python-docx>=1.1
